step-1: rename folder
